but: Move the editor function to a shared module
The editor function is in a shared module so that it doesn't have
to be reimplemented everywhere
