+++ crates/but/src/describe/mod.rs
@@ -4,7 +4,7 @@ use gitbutler_command_context::CommandContext;
 use gitbutler_oxidize::ObjectIdExt;
 use gitbutler_project::Project;
 
-use crate::id::CliId;
+use crate::{editor::get_text_from_editor_no_comments, id::CliId};
 
 pub(crate) fn describe_target(project: &Project, _json: bool, target: &str) -> Result<()> {
     let mut ctx = CommandContext::open(project, AppSettings::load_from_default_path_creating()?)?;
@@ -178,13 +178,6 @@ fn get_commit_message_from_editor(
     current_message: &str,
     changed_files: &[String],
 ) -> Result<String> {
-    // Get editor command
-    let editor = get_editor_command()?;
-
-    // Create temporary file with current message and file list
-    let temp_dir = std::env::temp_dir();
-    let temp_file = temp_dir.join(format!("but_commit_msg_{}", std::process::id()));
-
     // Generate commit message template with current message
     let mut template = String::new();
     template.push_str(current_message);
@@ -201,28 +194,8 @@ fn get_commit_message_from_editor(
     }
     template.push_str("#\n");
 
-    std::fs::write(&temp_file, template)?;
-
-    // Launch editor
-    let status = std::process::Command::new(&editor)
-        .arg(&temp_file)
-        .status()?;
-
-    if !status.success() {
-        anyhow::bail!("Editor exited with non-zero status");
-    }
-
     // Read the result and strip comments
-    let content = std::fs::read_to_string(&temp_file)?;
-    std::fs::remove_file(&temp_file).ok(); // Best effort cleanup
-
-    let message = content
-        .lines()
-        .filter(|line| !line.starts_with('#'))
-        .collect::<Vec<_>>()
-        .join("\n")
-        .trim()
-        .to_string();
+    let message = get_text_from_editor_no_comments("but_commit_msg", &template)?;
 
     if message.is_empty() {
         anyhow::bail!("Aborting due to empty commit message");
@@ -232,11 +205,6 @@ fn get_commit_message_from_editor(
 }
 
 fn get_branch_name_from_editor(current_name: &str) -> Result<String> {
-    let editor = get_editor_command()?;
-
-    let temp_dir = std::env::temp_dir();
-    let temp_file = temp_dir.join(format!("but_branch_name_{}", std::process::id()));
-
     let mut template = String::new();
     template.push_str(current_name);
     if !current_name.is_empty() && !current_name.ends_with('\n') {
@@ -246,56 +214,11 @@ fn get_branch_name_from_editor(current_name: &str) -> Result<String> {
     template.push_str("# with '#' will be ignored, and an empty name aborts the operation.\n");
     template.push_str("#\n");
 
-    std::fs::write(&temp_file, template)?;
-
-    let status = std::process::Command::new(&editor)
-        .arg(&temp_file)
-        .status()?;
-
-    if !status.success() {
-        anyhow::bail!("Editor exited with non-zero status");
-    }
-
-    let content = std::fs::read_to_string(&temp_file)?;
-    std::fs::remove_file(&temp_file).ok();
-
-    let branch_name = content
-        .lines()
-        .filter(|line| !line.starts_with('#'))
-        .collect::<Vec<_>>()
-        .join("\n")
-        .trim()
-        .to_string();
+    let branch_name = get_text_from_editor_no_comments("but_branch_name", &template)?;
 
     if branch_name.is_empty() {
         anyhow::bail!("Aborting due to empty branch name");
     }
 
     Ok(branch_name)
-}
-
-fn get_editor_command() -> Result<String> {
-    // Try $EDITOR first
-    if let Ok(editor) = std::env::var("EDITOR") {
-        return Ok(editor);
-    }
-
-    // Try git config core.editor
-    if let Ok(output) = std::process::Command::new("git")
-        .args(["config", "--get", "core.editor"])
-        .output()
-        && output.status.success()
-    {
-        let editor = String::from_utf8_lossy(&output.stdout).trim().to_string();
-        if !editor.is_empty() {
-            return Ok(editor);
-        }
-    }
-
-    // Fallback to platform defaults
-    #[cfg(windows)]
-    return Ok("notepad".to_string());
-
-    #[cfg(not(windows))]
-    return Ok("vi".to_string());
 }
